+++ internal/provider/provider.go
@@ -9,6 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 
 	cmlclient "github.com/rschmied/gocmlclient"
@@ -49,44 +50,43 @@ func (p *CML2Provider) Configure(ctx context.Context, req provider.ConfigureRequ
 	}
 
 	// check if provided auth configuration makes sense
-	if data.Token.Null &&
-		(data.Username.Null || data.Password.Null) {
+	if data.Token.IsNull() &&
+		(data.Username.IsNull() || data.Password.IsNull()) {
 		resp.Diagnostics.AddError(
 			"Required configuration missing",
 			fmt.Sprintf("null check: either username and password or a token must be provided %T", p),
 		)
 	}
 
-	if len(data.Token.Value) == 0 &&
-		(len(data.Username.Value) == 0 || len(data.Password.Value) == 0) {
+	if len(data.Token.ValueString()) == 0 &&
+		(len(data.Username.ValueString()) == 0 || len(data.Password.ValueString()) == 0) {
 		resp.Diagnostics.AddError(
 			"Required configuration missing",
 			fmt.Sprintf("value check: either username and password or a token must be provided %T", p),
 		)
 	}
 
-	if len(data.Token.Value) > 0 && len(data.Username.Value) > 0 {
+	if len(data.Token.ValueString()) > 0 && len(data.Username.ValueString()) > 0 {
 		resp.Diagnostics.AddWarning(
 			"Conflicting configuration",
 			"both token and username / password were provided")
 	}
 
 	// an address must be specified
-	if len(data.Address.Value) == 0 {
+	if len(data.Address.ValueString()) == 0 {
 		resp.Diagnostics.AddError(
 			"Required configuration missing",
 			fmt.Sprintf("a server address must be configured to use %T", p),
 		)
 	}
-	if data.SkipVerify.Null {
+	if data.SkipVerify.IsNull() {
 		tflog.Warn(ctx, "unspecified certificate verification, will verify")
-		data.SkipVerify.Value = false
+		data.SkipVerify = types.BoolValue(false)
 	}
 
-	if data.UseCache.Null {
-		data.SkipVerify.Value = false
-	}
-	if data.UseCache.Value {
+	if data.UseCache.IsNull() {
+		data.UseCache = types.BoolValue(false)
+	} else if data.UseCache.ValueBool() {
 		resp.Diagnostics.AddWarning(
 			"Experimental feature enabled",
 			"\"use_cache\" is considered experimental and may not work as expected; use with care",
@@ -95,19 +95,22 @@ func (p *CML2Provider) Configure(ctx context.Context, req provider.ConfigureRequ
 
 	// create a new CML2 client
 	client := cmlclient.NewClient(
-		data.Address.Value,
-		data.SkipVerify.Value,
-		data.UseCache.Value,
+		data.Address.ValueString(),
+		data.SkipVerify.ValueBool(),
+		data.UseCache.ValueBool(),
 	)
-	if len(data.Username.Value) > 0 {
-		client.SetUsernamePassword(data.Username.Value, data.Password.Value)
+	if len(data.Username.ValueString()) > 0 {
+		client.SetUsernamePassword(
+			data.Username.ValueString(),
+			data.Password.ValueString(),
+		)
 	}
-	if len(data.Token.Value) > 0 {
-		client.SetToken(data.Token.Value)
+	if len(data.Token.ValueString()) > 0 {
+		client.SetToken(data.Token.ValueString())
 	}
 
-	if len(data.CAcert.Value) > 0 {
-		err := client.SetCACert([]byte(data.CAcert.Value))
+	if len(data.CAcert.ValueString()) > 0 {
+		err := client.SetCACert([]byte(data.CAcert.ValueString()))
 		if err != nil {
 			resp.Diagnostics.AddError(
 				"Configuration issue",