+++ internal/provider/resource/lab/modify_plan.go
@@ -5,15 +5,20 @@ import (
 	"reflect"
 
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/rschmied/terraform-provider-cml2/internal/schema"
 )
 
 func (r *LabResource) ModifyPlan(ctx context.Context, req resource.ModifyPlanRequest, resp *resource.ModifyPlanResponse) {
 
-	var stateData, planData *schema.LabModel
+	var stateData, planData schema.LabModel
 
-	tflog.Info(ctx, "Resource Lab ModifyPlan")
+	tflog.Info(ctx, "Resource Lab MODIFYPLAN")
+
+	if req.Plan.Raw.IsNull() || req.State.Raw.IsNull() {
+		return
+	}
 
 	// Read Terraform plan data into the model
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &planData)...)
@@ -27,16 +32,11 @@ func (r *LabResource) ModifyPlan(ctx context.Context, req resource.ModifyPlanReq
 		return
 	}
 
-	// when deleting, there's no plan
-	if planData == nil {
-		return
-	}
-
 	// if state and plan are identical -> modified date has changed
 	// this gets auto-updated when we change something
 	if !reflect.DeepEqual(stateData, planData) {
-		planData.Modified.Unknown = true
+		planData.Modified = types.StringUnknown()
 	}
 	resp.Diagnostics.Append(resp.Plan.Set(ctx, planData)...)
-	tflog.Info(ctx, "Resource Lab ModifyPlan: done")
+	tflog.Info(ctx, "Resource Lab MODIFYPLAN: done")
 }