+++ internal/provider/resource/lifecycle/create.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
@@ -14,24 +15,31 @@ import (
 
 func (r *LabLifecycleResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	var (
-		data *schema.LabLifecycleModel
+		data schema.LabLifecycleModel
 		err  error
 	)
+
+	tflog.Info(ctx, "Resource Lifecycle CREATE")
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
+	// create a resource identifier
+	if data.ID.IsUnknown() {
+		data.ID = types.StringValue(uuid.New().String())
+	}
+
 	start := startData{
 		staging:  getStaging(ctx, req.Config, &resp.Diagnostics),
 		timeouts: getTimeouts(ctx, req.Config, &resp.Diagnostics),
-		wait:     data.Wait.Null || data.Wait.Value,
+		wait:     data.Wait.IsNull() || data.Wait.ValueBool(),
 	}
 
-	if data.ID.IsUnknown() {
+	if data.LabID.IsUnknown() {
 		tflog.Info(ctx, "Create: import")
-		start.lab, err = r.cfg.Client().LabImport(ctx, data.Topology.Value)
+		start.lab, err = r.cfg.Client().LabImport(ctx, data.Topology.ValueString())
 		if err != nil {
 			resp.Diagnostics.AddError(
 				CML2ErrorLabel,
@@ -39,9 +47,9 @@ func (r *LabLifecycleResource) Create(ctx context.Context, req resource.CreateRe
 			)
 			return
 		}
-		data.ID = types.String{Value: start.lab.ID}
+		data.LabID = types.StringValue(start.lab.ID)
 	} else {
-		start.lab, err = r.cfg.Client().LabGet(ctx, data.ID.Value, true)
+		start.lab, err = r.cfg.Client().LabGet(ctx, data.LabID.ValueString(), true)
 		if err != nil {
 			resp.Diagnostics.AddError(
 				CML2ErrorLabel,
@@ -52,13 +60,13 @@ func (r *LabLifecycleResource) Create(ctx context.Context, req resource.CreateRe
 	}
 
 	// inject the configurations into the nodes
-	r.injectConfigs(ctx, start.lab, data, &resp.Diagnostics)
+	r.injectConfigs(ctx, start.lab, &data, &resp.Diagnostics)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
 	// if unknown state or specifically "start" state, start the lab...
-	if data.State.Unknown || data.State.Value == cmlclient.LabStateStarted {
+	if data.State.IsUnknown() || data.State.ValueString() == cmlclient.LabStateStarted {
 		r.startNodes(ctx, &resp.Diagnostics, start)
 	}
 
@@ -72,11 +80,11 @@ func (r *LabLifecycleResource) Create(ctx context.Context, req resource.CreateRe
 		return
 	}
 
-	data.ID = types.String{Value: lab.ID}
-	data.State = types.String{Value: lab.State}
+	data.LabID = types.StringValue(lab.ID)
+	data.State = types.StringValue(lab.State)
 	data.Nodes = r.populateNodes(ctx, lab, &resp.Diagnostics)
-	data.Booted = types.Bool{Value: lab.Booted()}
+	data.Booted = types.BoolValue(lab.Booted())
 
-	resp.Diagnostics.Append(resp.State.Set(ctx, data)...)
-	tflog.Info(ctx, "Create: done")
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+	tflog.Info(ctx, "Resource Lifecycle CREATE: done")
 }