+++ internal/provider/resource/lifecycle/modify_plan.go
@@ -15,16 +15,14 @@ import (
 
 func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.ModifyPlanRequest, resp *resource.ModifyPlanResponse) {
 
-	var configData, planData, stateData *schema.LabLifecycleModel
+	var configData, planData, stateData schema.LabLifecycleModel
 
-	// Read Terraform plan data into the model
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &planData)...)
-	if resp.Diagnostics.HasError() {
+	tflog.Info(ctx, "Resource Lifecycle MODIFYPLAN")
+
+	// configuration data for the resource
+	if req.Config.Raw.IsNull() {
 		return
 	}
-
-	tflog.Info(ctx, "ModifyPlan")
-
 	resp.Diagnostics.Append(req.Config.Get(ctx, &configData)...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -39,40 +37,29 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 		}
 	}
 
+	// not much to do without a plan...
+	if req.Plan.Raw.IsNull() {
+		tflog.Error(ctx, "ModifyPlan: no plan exists...")
+		return
+	}
+
 	// get the planned state
-	resp.Diagnostics.Append(resp.Plan.Get(ctx, &planData)...)
+	resp.Diagnostics.Append(req.Plan.Get(ctx, &planData)...)
 	if resp.Diagnostics.HasError() {
 		tflog.Error(ctx, "ModifyPlan: plan has errors")
 		return
 	}
 
-	if planData == nil {
-		tflog.Error(ctx, "ModifyPlan: no plan exists...")
-		return
-	}
-
-	// if !noState && !configData.Elements.IsNull() {
-	// 	for idx := 0; idx < len(planData.Elements.Elems); idx++ {
-	// 		id := stateData.Elements.Elems[idx].(types.String)
-	// 		id.Null = false
-	// 		id.Unknown = false
-	// 		resp.Diagnostics.Append(tfsdk.ValueFrom(ctx, id, types.StringType, &planData.Elements.Elems[idx])...)
-	// 		if resp.Diagnostics.HasError() {
-	// 			return
-	// 		}
-	// 	}
-	// }
-
 	// check if we can transition to specified state
-	if planData.State.Value == cmlclient.LabStateStopped {
-		if !noState && stateData.State.Value == cmlclient.LabStateDefined {
+	if planData.State.ValueString() == cmlclient.LabStateStopped {
+		if !noState && stateData.State.ValueString() == cmlclient.LabStateDefined {
 			resp.Diagnostics.AddError(
 				CML2ErrorLabel,
 				"can't transition from DEFINED_ON_CORE to STOPPED",
 			)
 			return
 		}
-		if noState && planData.State.Value == cmlclient.LabStateStopped {
+		if noState && planData.State.ValueString() == cmlclient.LabStateStopped {
 			resp.Diagnostics.AddError(
 				CML2ErrorLabel,
 				"can't transition from no state to STOPPED",
@@ -81,9 +68,13 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 		}
 	}
 
+	if configData.State.IsNull() {
+		planData.State = types.StringValue("STARTED")
+	}
+
 	changeNeeded := false
 	if !noState {
-		changeNeeded = planData.State.Value != stateData.State.Value
+		changeNeeded = planData.State.ValueString() != stateData.State.ValueString()
 	}
 
 	if changeNeeded {
@@ -98,32 +89,44 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 
 		for id, node := range nodes {
 
+			planState := planData.State.ValueString()
+
 			// these all need to be re-read when state changes...  based on
 			// actual state change, these can be optimized to provide a better
 			// state diff -- but it works for now
-			node.State.Unknown = true
-			node.State.Null = false
 
-			node.DataVolume.Unknown = true
-			node.DataVolume.Null = false
+			// node.DataVolume = types.Int64Unknown()
+			// node.CPUs = types.Int64Unknown()
+			// node.RAM = types.Int64Unknown()
+			// node.BootDiskSize = types.Int64Unknown()
 
-			node.ComputeID.Unknown = true
-			node.ComputeID.Null = false
+			if planState != "STARTED" {
+				node.State = types.StringValue(planData.State.ValueString())
+			}
 
-			node.SerialDevices.Unknown = true
-			node.SerialDevices.Null = false
-
-			node.CPUs.Unknown = true
-			node.CPUs.Null = false
-
-			node.VNCkey.Unknown = true
-			node.VNCkey.Null = false
-
-			node.RAM.Unknown = true
-			node.RAM.Null = false
-
-			node.BootDiskSize.Unknown = true
-			node.BootDiskSize.Null = false
+			if planData.State.ValueString() == "DEFINED_ON_CORE" {
+				node.SerialDevices = types.ListNull(schema.SerialDevicesAttrType)
+				node.VNCkey = types.StringNull()
+				node.ComputeID = types.StringNull()
+				node.DataVolume = types.Int64Null()
+				node.CPUs = types.Int64Null()
+				node.RAM = types.Int64Null()
+				node.BootDiskSize = types.Int64Null()
+				node.State = types.StringValue("DEFINED_ON_CORE")
+			}
+			if planData.State.ValueString() == "STARTED" {
+				node.SerialDevices = types.ListUnknown(schema.SerialDevicesAttrType)
+				node.VNCkey = types.StringUnknown()
+				node.ComputeID = types.StringUnknown()
+				node.DataVolume = types.Int64Unknown()
+				node.CPUs = types.Int64Unknown()
+				node.RAM = types.Int64Unknown()
+				node.BootDiskSize = types.Int64Unknown()
+				node.State = types.StringUnknown()
+			}
+			if planData.State.ValueString() == "STOPPED" {
+				node.State = types.StringValue("STOPPED")
+			}
 
 			// This is a bit of a hack since the node def name is hard coded
 			// here.  what happens is that UMS nodes get the bridge name as the
@@ -131,8 +134,8 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 			// start, the configuration is set to the name of the bridge, like
 			// ums-b843d547-54.
 			// As an alternative, all configurations could be set to "Unknown"
-			if node.NodeDefinition.Value == "unmanaged_switch" {
-				node.Configuration.Unknown = true
+			if node.NodeDefinition.ValueString() == "unmanaged_switch" {
+				node.Configuration = types.StringUnknown()
 			}
 
 			var ifaces []schema.InterfaceModel
@@ -141,25 +144,48 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 				return
 			}
 
-			for idx := range ifaces {
-				ifaces[idx].IP4.Unknown = true
-				ifaces[idx].IP6.Unknown = true
-				// we know that when we wipe, the MAC is going to be null
-				if planData.State.Value == "DEFINED_ON_CORE" {
-					ifaces[idx].MACaddress.Unknown = false
-					ifaces[idx].MACaddress.Null = true
-				} else {
-					// MACaddresses won't change at state change if one was assigned
-					if ifaces[idx].MACaddress.Null {
-						ifaces[idx].MACaddress.Unknown = true
-						ifaces[idx].MACaddress.Null = false // why? oh, why!
-					}
-				}
-				ifaces[idx].State.Unknown = true
+			// planState := planData.State.ValueString()
+			// for idx := range ifaces {
+			// 	ifaces[idx].IP4 = types.ListUnknown(types.StringType)
+			// 	ifaces[idx].IP6 = types.ListUnknown(types.StringType)
+			// 	// we know that when we wipe, the MAC is going to be null
+			// 	if planState == "DEFINED_ON_CORE" {
+			// 		ifaces[idx].MACaddress = types.StringNull()
+			// 		ifaces[idx].IP4 = types.ListNull(types.StringType)
+			// 		ifaces[idx].IP6 = types.ListNull(types.StringType)
+			// 	} else {
+			// 		// MACaddresses won't change at state change if one was assigned
+			// 		if ifaces[idx].MACaddress.IsNull() {
+			// 			ifaces[idx].MACaddress = types.StringUnknown()
+			// 		}
+			// 	}
+			// 	// if planData.State.ValueString() == "STOPPED" {
+			// 	// 	ifaces[idx].IP4 = types.ListNull(types.StringType)
+			// 	// 	ifaces[idx].IP6 = types.ListNull(types.StringType)
+			// 	// }
+			// 	ifaces[idx].State = types.StringUnknown()
 
-				// iface := ifaces[idx]
-				// tflog.Info(ctx, fmt.Sprintf("mac: %v/%v", iface.MACaddress.Null, iface.MACaddress.Unknown))
-				// tflog.Info(ctx, fmt.Sprintf("ip4: %v/%v", iface.IP4.Null, iface.IP4.Unknown))
+			for idx := range ifaces {
+				if planState == "STARTED" {
+					ifaces[idx].IP4 = types.ListUnknown(types.StringType)
+					ifaces[idx].IP6 = types.ListUnknown(types.StringType)
+					// MACaddresses won't change at state change if one was assigned
+					if ifaces[idx].MACaddress.IsNull() {
+						ifaces[idx].MACaddress = types.StringUnknown()
+					}
+					ifaces[idx].State = types.StringUnknown()
+				}
+				if planState == "DEFINED_ON_CORE" || planState == "STOPPED" {
+					ifaces[idx].IP4 = types.ListNull(types.StringType)
+					ifaces[idx].IP6 = types.ListNull(types.StringType)
+				}
+				if planState == "DEFINED_ON_CORE" {
+					ifaces[idx].MACaddress = types.StringNull()
+					ifaces[idx].State = types.StringValue("DEFINED_ON_CORE")
+				}
+				if planState == "STOPPED" {
+					ifaces[idx].State = types.StringValue("STOPPED")
+				}
 			}
 
 			resp.Diagnostics.Append(
@@ -188,10 +214,15 @@ func (r *LabLifecycleResource) ModifyPlan(ctx context.Context, req resource.Modi
 			return
 		}
 
-		// booted state of lab is unknown at this point
-		planData.Booted.Unknown = true
+		// booted state of lab is unknown if the plan is to start
+		if planData.State.ValueString() == "STARTED" {
+			planData.Booted = types.BoolUnknown()
+		} else {
+			planData.Booted = types.BoolValue(false)
+		}
 	}
 
-	resp.Diagnostics.Append(resp.Plan.Set(ctx, planData)...)
-	tflog.Info(ctx, "ModifyPlan: done")
+	resp.Diagnostics.Append(resp.Plan.Set(ctx, &planData)...)
+
+	tflog.Info(ctx, "Resource Lifecycle MODIFYPLAN: done")
 }